♻️ refactor(openapi): remove HttpxClientConfig from utils
- eliminate HttpxClientConfig class from openapi/utils.py
- update imports to reflect relocation of HttpxClientConfig
